requirements.txt
PyQt6>=6.0
python-vlc>=3.0
